src/backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
branca==0.8.1
celery==5.4.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
folium==0.19.5
idna==3.10
Jinja2==3.1.6
kombu==5.5.1
MarkupSafe==3.0.2
numpy==2.2.4
pillow==11.1.0
polyline==2.0.2
prompt_toolkit==3.0.50
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
reportlab==4.3.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
xyzservices==2025.1.0
